refactor: consolidate notifier layout and move supervisor to serve wiring
Merge notifier/runtime helper files, colocate flat keys and API view logic
in notifier, drop NotifierBackground from server in favor of explicit pull
supervisor startup in cli/serve, and reuse runtime.CloneAnyMap in bot.
