+++ cli/serve/serve.go
@@ -37,8 +37,6 @@ import (
 	internalonboard "csgclaw/internal/onboard"
 	agentruntime "csgclaw/internal/runtime"
 	runtimecodex "csgclaw/internal/runtime/codex"
-	runtimenotifier "csgclaw/internal/runtime/notifier"
-	notifierpull "csgclaw/internal/runtime/notifier/pull"
 	"csgclaw/internal/sandboxproviders"
 	"csgclaw/internal/server"
 	"csgclaw/internal/upgrade"
@@ -466,7 +464,7 @@ func startServerWithConfigPath(ctx context.Context, run *command.Context, cfg co
 	if err != nil {
 		return err
 	}
-	return RunServer(server.Options{
+	serverOpts := server.Options{
 		ListenAddr:  cfg.Server.ListenAddr,
 		Service:     svc,
 		Hub:         hubSvc,
@@ -481,38 +479,36 @@ func startServerWithConfigPath(ctx context.Context, run *command.Context, cfg co
 		AccessToken: cfg.Server.AccessToken,
 		NoAuth:      cfg.Server.NoAuth,
 		Context:     ctx,
-		NotifierBackground: func(bgCtx context.Context, agentSvc *agent.Service, deliver runtimenotifier.Fanouter) {
-			if agentSvc == nil || deliver == nil {
-				return
-			}
-			notifierpull.NewSupervisor(agentSvc, deliver).Run(bgCtx)
-		},
-		OnReady: func() {
-			if output != "json" && run != nil {
-				go func() {
-					if err := WaitForHealthy(apiURL, 5*time.Second); err != nil {
-						fmt.Fprintln(run.Stdout, "Open this URL in your browser after startup.")
-						return
-					}
-					if err := OpenBrowser(imURL); err != nil {
-						fmt.Fprintln(run.Stdout, "Open this URL in your browser after startup.")
-					} else {
-						fmt.Fprintln(run.Stdout, "Opened this URL in your browser.")
-					}
-				}()
-			}
+	}
+	handler := server.NewHandler(serverOpts)
+	serverOpts.Handler = handler
+	go runtimewiring.RunNotifierPullSupervisor(ctx, svc, handler)
+	serverOpts.OnReady = func() {
+		if output != "json" && run != nil {
 			go func() {
-				if err := StartConfiguredAgents(ctx, svc); err != nil {
-					slog.Warn("some configured agents failed to start", "error", err)
+				if err := WaitForHealthy(apiURL, 5*time.Second); err != nil {
+					fmt.Fprintln(run.Stdout, "Open this URL in your browser after startup.")
+					return
 				}
-				if codexBridgeMgr != nil {
-					if err := codexBridgeMgr.Start(ctx); err != nil {
-						slog.Warn("some codex bridges failed to start", "error", err)
-					}
+				if err := OpenBrowser(imURL); err != nil {
+					fmt.Fprintln(run.Stdout, "Open this URL in your browser after startup.")
+				} else {
+					fmt.Fprintln(run.Stdout, "Opened this URL in your browser.")
 				}
 			}()
-		},
-	})
+		}
+		go func() {
+			if err := StartConfiguredAgents(ctx, svc); err != nil {
+				slog.Warn("some configured agents failed to start", "error", err)
+			}
+			if codexBridgeMgr != nil {
+				if err := codexBridgeMgr.Start(ctx); err != nil {
+					slog.Warn("some codex bridges failed to start", "error", err)
+				}
+			}
+		}()
+	}
+	return RunServer(serverOpts)
 }
 
 func configureFeishuService(feishuSvc *feishu.Service, svc *agent.Service) {
