+++ internal/app/runtimewiring/notifier.go
@@ -1,10 +1,12 @@
 package runtimewiring
 
 import (
+	"context"
 	"fmt"
 
 	"csgclaw/internal/agent"
 	runtimenotifier "csgclaw/internal/runtime/notifier"
+	notifierpull "csgclaw/internal/runtime/notifier/pull"
 )
 
 func WithNotifierRuntime() agent.ServiceOption {
@@ -14,4 +16,13 @@ func WithNotifierRuntime() agent.ServiceOption {
 		}
 		return agent.WithRuntime(runtimenotifier.NewAgentRuntime())(s)
 	}
+}
+
+// RunNotifierPullSupervisor blocks until ctx is cancelled, reconciling per-agent remote_pull loops.
+// Start it from the composition layer (e.g. cli/serve) with go RunNotifierPullSupervisor(ctx, ...).
+func RunNotifierPullSupervisor(ctx context.Context, agents *agent.Service, deliver runtimenotifier.Fanouter) {
+	if agents == nil || deliver == nil {
+		return
+	}
+	notifierpull.NewSupervisor(agents, deliver).Run(ctx)
 }
