+++ internal/runtime/notifier/subscription.go
@@ -4,6 +4,7 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"fmt"
+	"strings"
 	"time"
 )
 
@@ -24,6 +25,23 @@ func EnsurePullRemoteSubscriptionInRequestOptions(ro map[string]any) map[string]
 	EnsurePullRemoteSubscriptionInNotifierDetails(m)
 	ro["notifier"] = m
 	return ro
+}
+
+// EnsurePullRemoteSubscriptionInNotifierDetails sets remote_subscription_id on a flat notifier map
+// when delivery_mode is remote_pull and the id is empty. Mutates nd in place.
+func EnsurePullRemoteSubscriptionInNotifierDetails(nd map[string]any) map[string]any {
+	if len(nd) == 0 {
+		return nd
+	}
+	cfg := ParseNotifierDetails(nd)
+	if cfg.normalizedDeliveryMode() != DeliveryRemotePull {
+		return nd
+	}
+	if strings.TrimSpace(cfg.RemoteSubscriptionID) != "" {
+		return nd
+	}
+	nd["remote_subscription_id"] = newPullSubscriptionID()
+	return nd
 }
 
 func newPullSubscriptionID() string {
